cloud_functions/basic_get_function/requirements.txt
google-cloud-firestore==1.3.0
flask==1.1.2